# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "MusicBox API Documentation",
        "description": "API Documentation for MusicBox",
        "contact": {
            "email": "your-email@domain.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api",
            "description": "Local API Server"
        }
    ],
    "paths": {
        "/albums": {
            "get": {
                "tags": [
                    "Albums"
                ],
                "summary": "Get list of albums",
                "operationId": "1da83b13e3f02c4fee3c24e70098d603",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "albums": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "title": {
                                                        "type": "string",
                                                        "example": "Album Title"
                                                    },
                                                    "genre": {
                                                        "type": "string",
                                                        "example": "Rock"
                                                    },
                                                    "release_date": {
                                                        "type": "string",
                                                        "format": "date",
                                                        "example": "2023-09-30"
                                                    },
                                                    "artist_id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Albums Fetched Successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Albums"
                ],
                "summary": "Create a new album",
                "operationId": "e2953334b8702ac4d6cc93c63b7440ef",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "title",
                                    "genre",
                                    "release_date",
                                    "artist_id"
                                ],
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "example": "New Album"
                                    },
                                    "genre": {
                                        "type": "string",
                                        "example": "Rock"
                                    },
                                    "release_date": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2023-09-30"
                                    },
                                    "artist_id": {
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Album created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "album": {
                                            "type": "object"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Album Created Successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/albums/{id}": {
            "get": {
                "tags": [
                    "Albums"
                ],
                "summary": "Get album by ID with artist and songs",
                "operationId": "6a68209549d4d66cd13500611af64336",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "album": {
                                            "type": "object"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Album Fetched Successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Albums"
                ],
                "summary": "Update album",
                "operationId": "9d45cc434d36c93ef067bbd2be12a53c",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "title",
                                    "genre",
                                    "release_date",
                                    "artist_id"
                                ],
                                "properties": {
                                    "title": {
                                        "type": "string"
                                    },
                                    "genre": {
                                        "type": "string"
                                    },
                                    "release_date": {
                                        "type": "string",
                                        "format": "date"
                                    },
                                    "artist_id": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "album": {
                                            "type": "object"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Album Updated Successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Albums"
                ],
                "summary": "Delete album",
                "operationId": "b148d53b889a5c777b217bb73821307e",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Album Deleted Successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    },
...
}